Fixed missing \n from payload
Using mosquitto_pub -l we need a \n at the end of our payload, or it
will be ignored, and mosquitto_pub will never exit.

Set is not a list so don't treat it like one; make a list for it.

Moved code for config to be more global; read it only at startup.

Removed some debugging code