Changed code to not use hermes_python.publish_continue_session
That code wasn't working properly, and I believe it was causing the app
to fail.  So instead, we use a real Lock (instead of relying on a
boolean check to be fast enough), and if it's locked, we just return an
error.  Since we don't really care if the injection starts while we're
checking the list, we don't maintain the lock, and release it immediately.
We keep the lock when adding items, though.